refactor: move skills to lib/zero/
Move skills module (3 source files) from src/lib/skills/ to
src/lib/zero/skills/. This module manages skill sync and already
imports from zero/seed-skills, making it a natural fit for the
zero/ layer.

Closes #7731
