+++ turbo/apps/web/app/api/cron/sync-skills/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { initServices } from "../../../../src/lib/init-services";
-import { syncSkills } from "../../../../src/lib/skills/sync-skills";
+import { syncSkills } from "../../../../src/lib/zero/skills/sync-skills";
 import { logger } from "../../../../src/lib/logger";
 import { env } from "../../../../src/env";
